+++ models/sipnet/R/model2netcdf.SIPNET.R
@@ -65,42 +65,42 @@ model2netcdf.SIPNET <- function(outdir, sitelat, sitelon, start_date, end_date, 
     bounds <- array(data=NA, dim=c(length(tvals),2))
     bounds[,1] <- tvals-1
     bounds[,2] <- bounds[,1]+dayfrac
-    #head(bounds); tail(bounds)
     
     ## Setup outputs for netCDF file in appropriate units
     output       <- list()
-    output[[1]]  <- (sub.sipnet.output$gpp * 0.001) / timestep.s  # GPP in kgC/m2/s
+    output[[1]]  <- c(rbind(bounds[,1], bounds[,2]))
+    output[[2]]  <- (sub.sipnet.output$gpp * 0.001) / timestep.s  # GPP in kgC/m2/s
     ## output[[2]] <- (sub.sipnet.output$npp*0.001) / timestep.s # NPP in kgC/m2/s. Internal SIPNET
     ## calculation
-    output[[2]]  <- (sub.sipnet.output$gpp * 0.001) / timestep.s - ((sub.sipnet.output$rAboveground *
+    output[[3]]  <- (sub.sipnet.output$gpp * 0.001) / timestep.s - ((sub.sipnet.output$rAboveground *
                                                                        0.001) / timestep.s + (sub.sipnet.output$rRoot * 0.001) / timestep.s)  # NPP in kgC/m2/s. Post SIPNET calculation
-    output[[3]]  <- (sub.sipnet.output$rtot * 0.001) / timestep.s  # Total Respiration in kgC/m2/s
-    output[[4]]  <- (sub.sipnet.output$rAboveground * 0.001) / timestep.s + (sub.sipnet.output$rRoot *
+    output[[4]]  <- (sub.sipnet.output$rtot * 0.001) / timestep.s  # Total Respiration in kgC/m2/s
+    output[[5]]  <- (sub.sipnet.output$rAboveground * 0.001) / timestep.s + (sub.sipnet.output$rRoot *
                                                                                0.001) / timestep.s  # Autotrophic Respiration in kgC/m2/s
-    output[[5]]  <- ((sub.sipnet.output$rSoil - sub.sipnet.output$rRoot) * 0.001) / timestep.s  # Heterotrophic Respiration in kgC/m2/s
-    output[[6]]  <- (sub.sipnet.output$rSoil * 0.001) / timestep.s  # Soil Respiration in kgC/m2/s
-    output[[7]]  <- (sub.sipnet.output$nee * 0.001) / timestep.s  # NEE in kgC/m2/s
+    output[[6]]  <- ((sub.sipnet.output$rSoil - sub.sipnet.output$rRoot) * 0.001) / timestep.s  # Heterotrophic Respiration in kgC/m2/s
+    output[[7]]  <- (sub.sipnet.output$rSoil * 0.001) / timestep.s  # Soil Respiration in kgC/m2/s
+    output[[8]]  <- (sub.sipnet.output$nee * 0.001) / timestep.s  # NEE in kgC/m2/s
     # output[[7]] <- rep(-999,sipnet.output.dims[1]) # CarbPools
-    output[[8]] <- (sub.sipnet.output$plantWoodC * 0.001)  # Above ground wood kgC/m2
-    output[[9]] <- (sub.sipnet.output$plantLeafC * 0.001)  # Leaf C kgC/m2
-    output[[10]] <- (sub.sipnet.output$plantWoodC * 0.001) + (sub.sipnet.output$plantLeafC * 0.001) + 
+    output[[9]] <- (sub.sipnet.output$plantWoodC * 0.001)  # Above ground wood kgC/m2
+    output[[10]] <- (sub.sipnet.output$plantLeafC * 0.001)  # Leaf C kgC/m2
+    output[[11]] <- (sub.sipnet.output$plantWoodC * 0.001) + (sub.sipnet.output$plantLeafC * 0.001) + 
       (sub.sipnet.output$coarseRootC + sub.sipnet.output$fineRootC) * 0.001 # Total living C kgC/m2
-    output[[11]] <- (sub.sipnet.output$soil * 0.001) + (sub.sipnet.output$litter * 0.001)  # Total soil C kgC/m2
+    output[[12]] <- (sub.sipnet.output$soil * 0.001) + (sub.sipnet.output$litter * 0.001)  # Total soil C kgC/m2
     if (revision == "r136") {
-      output[[12]] <- (sub.sipnet.output$evapotranspiration * 10 * PEcAn.data.atmosphere::get.lv()) / timestep.s  # Qle W/m2
+      output[[13]] <- (sub.sipnet.output$evapotranspiration * 10 * PEcAn.data.atmosphere::get.lv()) / timestep.s  # Qle W/m2
     } else {
       ## *** NOTE : npp in the sipnet output file is actually evapotranspiration, this is due to a bug in sipnet.c : ***
       ## *** it says "npp" in the header (written by L774) but the values being written are trackers.evapotranspiration (L806) ***
       ## evapotranspiration in SIPNET is cm^3 water per cm^2 of area, to convert it to latent heat units W/m2 multiply with :
       ## 0.01 (cm2m) * 1000 (water density, kg m-3) * latent heat of vaporization (J kg-1)
       ## latent heat of vaporization is not constant and it varies slightly with temperature, get.lv() returns 2.5e6 J kg-1 by default
-      output[[12]] <- (sub.sipnet.output$npp * 10 * PEcAn.data.atmosphere::get.lv()) / timestep.s  # Qle W/m2
+      output[[13]] <- (sub.sipnet.output$npp * 10 * PEcAn.data.atmosphere::get.lv()) / timestep.s  # Qle W/m2
     }
-    output[[13]] <- (sub.sipnet.output$fluxestranspiration * 10) / timestep.s  # Transpiration kgW/m2/s
-    output[[14]] <- (sub.sipnet.output$soilWater * 10)  # Soil moisture kgW/m2
-    output[[15]] <- (sub.sipnet.output$soilWetnessFrac)  # Fractional soil wetness
-    output[[16]] <- (sub.sipnet.output$snow * 10)  # SWE
-    output[[17]] <- sub.sipnet.output$litter * 0.001  ## litter kgC/m2
+    output[[14]] <- (sub.sipnet.output$fluxestranspiration * 10) / timestep.s  # Transpiration kgW/m2/s
+    output[[15]] <- (sub.sipnet.output$soilWater * 10)  # Soil moisture kgW/m2
+    output[[16]] <- (sub.sipnet.output$soilWetnessFrac)  # Fractional soil wetness
+    output[[17]] <- (sub.sipnet.output$snow * 10)  # SWE
+    output[[18]] <- sub.sipnet.output$litter * 0.001  ## litter kgC/m2
 
     #calculate LAI for standard output
     param <- read.table(file.path(gsub(pattern = "/out/",
@@ -109,15 +109,14 @@ model2netcdf.SIPNET <- function(outdir, sitelat, sitelon, start_date, end_date, 
     id <- which(param[, 1] == "leafCSpWt")
     leafC <- 0.48
     SLA <- 1000 * leafC / param[id, 2] #SLA, m2/kgC
-    output[[18]] <- output[[9]] * SLA # LAI
+    output[[19]] <- output[[9]] * SLA # LAI
 
-    output[[19]] <- sub.sipnet.output$fineRootC   * 0.001  ## fine_root_carbon_content kgC/m2
-    output[[20]] <- sub.sipnet.output$coarseRootC * 0.001  ## coarse_root_carbon_content kgC/m2
-    output[[21]] <- (sub.sipnet.output$woodCreation * 0.001) / 86400 ## kgC/m2/s - this is daily in SIPNET
-    output[[22]] <- (sub.sipnet.output$plantWoodC + sub.sipnet.output$plantLeafC) * 0.001 # Total aboveground biomass kgC/m2
+    output[[20]] <- sub.sipnet.output$fineRootC   * 0.001  ## fine_root_carbon_content kgC/m2
+    output[[21]] <- sub.sipnet.output$coarseRootC * 0.001  ## coarse_root_carbon_content kgC/m2
+    output[[22]] <- (sub.sipnet.output$woodCreation * 0.001) / 86400 ## kgC/m2/s - this is daily in SIPNET
+    output[[23]] <- (sub.sipnet.output$plantWoodC + sub.sipnet.output$plantLeafC) * 0.001 # Total aboveground biomass kgC/m2
     
     # ******************** Declare netCDF variables ********************#
-
     t <- ncdf4::ncdim_def(name = "time",
                    longname = "time",
                    units = paste0("days since ", y, "-01-01 00:00:00"),
@@ -127,10 +126,11 @@ model2netcdf.SIPNET <- function(outdir, sitelat, sitelon, start_date, end_date, 
     lat <- ncdf4::ncdim_def("lat", "degrees_north", vals = as.numeric(sitelat), longname = "station_latitude")
     lon <- ncdf4::ncdim_def("lon", "degrees_east", vals = as.numeric(sitelon), longname = "station_longitude")
     dims <- list(lon = lon, lat = lat, time = t)
-    nb <- ncdf4::ncdim_def(name = "nb",vals =2, units='')
-    bdims <- list(time=t,nb)
-    nc_time_bounds <- ncdf4::ncvar_def(name="time_bounds", units=paste0("days since ", y, "-01-01 00:00:00"), 
-                                       longname = "time_bounds", dim=bdims, prec = "double")
+    time_interval <- ncdf4::ncdim_def(name = "time_interval", longname = "output time interval endpoints", 
+                                      vals = as.integer(1:2), units='')
+    #nc_time_bounds <- ncdf4::ncvar_def(name="time_bounds", units=paste0("days since ", y, "-01-01 00:00:00"), 
+    #                                   longname = "output time interval endpoints", dim=list(time_interval, time = t), 
+    #                                   prec = "float")
     
     ## ***** Need to dynamically update the UTC offset here *****
 
@@ -141,37 +141,39 @@ model2netcdf.SIPNET <- function(outdir, sitelat, sitelon, start_date, end_date, 
 
     mstmipvar <- PEcAn.utils::mstmipvar
     nc_var <- list()
-    nc_var[[1]]  <- PEcAn.utils::to_ncvar("GPP", dims)
-    nc_var[[2]]  <- PEcAn.utils::to_ncvar("NPP", dims)
-    nc_var[[3]]  <- PEcAn.utils::to_ncvar("TotalResp", dims)
-    nc_var[[4]]  <- PEcAn.utils::to_ncvar("AutoResp", dims)
-    nc_var[[5]]  <- PEcAn.utils::to_ncvar("HeteroResp", dims)
-    nc_var[[6]]  <- ncdf4::ncvar_def("SoilResp", units = "kg C m-2 s-1", dim = list(lon, lat, t), missval = -999,
+    nc_var[[1]] <- ncdf4::ncvar_def(name="time_bounds", units=paste0("days since ", y, "-01-01 00:00:00"), 
+                        longname = "output time bounds", dim=list(time_interval,time = t), prec = "double")
+    nc_var[[2]]  <- PEcAn.utils::to_ncvar("GPP", dims)
+    nc_var[[3]]  <- PEcAn.utils::to_ncvar("NPP", dims)
+    nc_var[[4]]  <- PEcAn.utils::to_ncvar("TotalResp", dims)
+    nc_var[[5]]  <- PEcAn.utils::to_ncvar("AutoResp", dims)
+    nc_var[[6]]  <- PEcAn.utils::to_ncvar("HeteroResp", dims)
+    nc_var[[7]]  <- ncdf4::ncvar_def("SoilResp", units = "kg C m-2 s-1", dim = list(lon, lat, t), missval = -999,
                           longname = "Soil Respiration") #need to figure out standard variable for this output
-    nc_var[[7]]  <- PEcAn.utils::to_ncvar("NEE", dims)
+    nc_var[[8]]  <- PEcAn.utils::to_ncvar("NEE", dims)
     # nc_var[[7]] <- mstmipvar('CarbPools', lat, lon, t, NA)
-    nc_var[[8]] <- PEcAn.utils::to_ncvar("AbvGrndWood", dims)
-    nc_var[[9]] <- PEcAn.utils::to_ncvar("leaf_carbon_content", dims)
-    nc_var[[10]] <- PEcAn.utils::to_ncvar("TotLivBiom", dims)
-    nc_var[[11]] <- PEcAn.utils::to_ncvar("TotSoilCarb", dims)
-    nc_var[[12]] <- PEcAn.utils::to_ncvar("Qle", dims)
-    nc_var[[13]] <- PEcAn.utils::to_ncvar("Transp", dims)
-    nc_var[[14]] <- PEcAn.utils::to_ncvar("SoilMoist", dims)
-    nc_var[[15]] <- PEcAn.utils::to_ncvar("SoilMoistFrac", dims)
-    nc_var[[16]] <- PEcAn.utils::to_ncvar("SWE", dims)
-    nc_var[[17]] <- PEcAn.utils::to_ncvar("litter_carbon_content", dims)
-    nc_var[[18]] <- PEcAn.utils::to_ncvar("LAI", dims)
-    nc_var[[19]] <- PEcAn.utils::to_ncvar("fine_root_carbon_content", dims)
-    nc_var[[20]] <- PEcAn.utils::to_ncvar("coarse_root_carbon_content", dims)
-    nc_var[[21]] <- ncdf4::ncvar_def("GWBI", units = "kg C m-2", dim = list(lon, lat, t), missval = -999,
+    nc_var[[9]] <- PEcAn.utils::to_ncvar("AbvGrndWood", dims)
+    nc_var[[10]] <- PEcAn.utils::to_ncvar("leaf_carbon_content", dims)
+    nc_var[[11]] <- PEcAn.utils::to_ncvar("TotLivBiom", dims)
+    nc_var[[12]] <- PEcAn.utils::to_ncvar("TotSoilCarb", dims)
+    nc_var[[13]] <- PEcAn.utils::to_ncvar("Qle", dims)
+    nc_var[[14]] <- PEcAn.utils::to_ncvar("Transp", dims)
+    nc_var[[15]] <- PEcAn.utils::to_ncvar("SoilMoist", dims)
+    nc_var[[16]] <- PEcAn.utils::to_ncvar("SoilMoistFrac", dims)
+    nc_var[[17]] <- PEcAn.utils::to_ncvar("SWE", dims)
+    nc_var[[18]] <- PEcAn.utils::to_ncvar("litter_carbon_content", dims)
+    nc_var[[19]] <- PEcAn.utils::to_ncvar("LAI", dims)
+    nc_var[[20]] <- PEcAn.utils::to_ncvar("fine_root_carbon_content", dims)
+    nc_var[[21]] <- PEcAn.utils::to_ncvar("coarse_root_carbon_content", dims)
+    nc_var[[22]] <- ncdf4::ncvar_def("GWBI", units = "kg C m-2", dim = list(lon, lat, t), missval = -999,
                                      longname = "Gross Woody Biomass Increment")
-    nc_var[[22]] <- ncdf4::ncvar_def("AGB", units = "kg C m-2", dim = list(lon, lat, t), missval = -999,
+    nc_var[[23]] <- ncdf4::ncvar_def("AGB", units = "kg C m-2", dim = list(lon, lat, t), missval = -999,
                                      longname = "Total aboveground biomass")
-
     # ******************** Declare netCDF variables ********************#
 
     ### Output netCDF data
     nc      <- ncdf4::nc_create(file.path(outdir, paste(y, "nc", sep = ".")), nc_var)
+    ncdf4::ncatt_put(nc, "time", "bounds", "time_bounds", prec=NA)
     varfile <- file(file.path(outdir, paste(y, "nc", "var", sep = ".")), "w")
     for (i in seq_along(nc_var)) {
       # print(i)
@@ -181,12 +183,17 @@ model2netcdf.SIPNET <- function(outdir, sitelat, sitelon, start_date, end_date, 
     close(varfile)
     ncdf4::nc_close(nc)
     
-    nc <- ncdf4::nc_open(file.path(outdir, paste(y, "nc", sep = ".")), write = TRUE)
-    ncdf4::ncatt_put(nc, "time", "bounds", "time_bounds", prec=NA, verbose=FALSE,
-               definemode=FALSE )
+    #nc <- ncdf4::nc_open(file.path(outdir, paste(y, "nc", sep = ".")), write = TRUE)
+    
+    #ncdf4::ncatt_put(nc, "time", "bounds", "time_bounds", prec=NA, verbose=FALSE,
+    #           definemode=FALSE )
+    ##ncdf4::ncatt_put(nc, "time", "time_bounds", bounds, prec="double", verbose=FALSE,
+    ##          definemode=FALSE )
+    #ncdf4::ncatt_put(nc, "time", "time_bounds", nc_time_bounds, prec="double", verbose=FALSE,
+    #                 definemode=FALSE )
     #ncdf4::ncvar_put(nc, nc_time_bounds, bounds)
-    ncdf4::ncvar_add( nc, nc_time_bounds, verbose=FALSE, indefine=FALSE )
-    ncdf4::nc_close(nc)
+    #ncdf4::ncvar_add( nc, nc_time_bounds, verbose=FALSE, indefine=FALSE )
+    #ncdf4::nc_close(nc)
     
   }  ### End of year loop